Add files via upload
Support Double as input for Graphite Client

+++ Graphite/GraphiteTcpClient.cs
@@ -20,7 +20,7 @@ namespace Graphite
             _tcpClient = new TcpClient(Hostname, Port);
         }
 
-        public void Send(string path, int value, DateTime timeStamp)
+        public void Send(string path, double value, DateTime timeStamp)
         {
             try
             {
